refactor: split tracedecay.rs indexing and file-scan into submodules
